+++ packages/services/guard/CHANGELOG.md
@@ -1,5 +1,107 @@
 # @0xsequence/guard
 
+## 3.0.0-beta.17
+
+### Patch Changes
+
+- New chains, relayer rpc fix
+
+## 3.0.0-beta.16
+
+### Patch Changes
+
+- ethauth support
+
+## 3.0.0-beta.15
+
+### Patch Changes
+
+- New chains, minor fixes
+
+## 3.0.0-beta.14
+
+### Patch Changes
+
+- Relayer fee options fix
+
+## 3.0.0-beta.13
+
+### Patch Changes
+
+- Userdata service updates
+
+## 3.0.0-beta.12
+
+### Patch Changes
+
+- Beta release with dapp connector fixes
+
+## 3.0.0-beta.11
+
+### Patch Changes
+
+- 3.0.0 beta
+
+## 3.0.0-beta.10
+
+### Patch Changes
+
+- dapp-client updates
+
+## 3.0.0-beta.9
+
+### Patch Changes
+
+- dapp client updates for EOA login
+
+## 3.0.0-beta.8
+
+### Patch Changes
+
+- Apple auth fixes
+
+## 3.0.0-beta.7
+
+### Patch Changes
+
+- Apple auth fix
+
+## 3.0.0-beta.6
+
+### Patch Changes
+
+- Fix signer 404 error, minor fixes
+
+## 3.0.0-beta.5
+
+### Patch Changes
+
+- Beta release for v3
+
+## 3.0.0-beta.4
+
+### Patch Changes
+
+- RC5 upgrade
+
+## 3.0.0-beta.3
+
+### Patch Changes
+
+- 3.0.0-beta.3 with fixes
+
+## 3.0.0-beta.2
+
+### Patch Changes
+
+- 3.0.0-beta.2 with identity instrument updates
+
+## 3.0.0-beta.1
+
+### Patch Changes
+
+- 3.0.0-beta.1
+
 ## 2.3.8
 
 ### Patch Changes
@@ -2302,7 +2404,6 @@
 - relayer: fix Relayer.wait() interface
 
   The interface for calling Relayer.wait() has changed. Instead of a single optional ill-defined timeout/delay parameter, there are three optional parameters, in order:
-
   - timeout: the maximum time to wait for the transaction receipt
   - delay: the polling interval, i.e. the time to wait between requests
   - maxFails: the maximum number of hard failures to tolerate before giving up